fix compiler -Wint-to-pointer-cast warnings at line 505/517, on casting params_to_pixels to XtPointer; no current xterm equivalent

+++ cxterm/scrollbar.c
@@ -466,9 +466,9 @@ static int specialcmplowerwiths (char *a, char *b)
     return 0;
 }
 
-static int params_to_pixels (TScreen *screen, String *params, int n)
+static long params_to_pixels (TScreen *screen, String *params, int n)
 {
-    register int mult = 1;
+    register long mult = 1;
     register char *s;
 
     switch (n > 2 ? 2 : n) {
